+++ hutool-extra/src/main/java/cn/hutool/extra/mail/MailUtil.java
@@ -345,7 +345,7 @@ public class MailUtil {
 	 * @since 4.6.3
 	 */
 	public static String send(MailAccount mailAccount, Collection<String> tos, Collection<String> ccs, Collection<String> bccs, String subject, String content, Map<String, InputStream> imageMap,
-	                          boolean isHtml, File... files) {
+							  boolean isHtml, File... files) {
 		return send(mailAccount, false, tos, ccs, bccs, subject, content, imageMap, isHtml, files);
 	}
 
@@ -357,7 +357,7 @@ public class MailUtil {
 	 * @return {@link Session}
 	 * @since 5.5.7
 	 */
-	public static Session getSession(MailAccount mailAccount, boolean isSingleton){
+	public static Session getSession(MailAccount mailAccount, boolean isSingleton) {
 		Authenticator authenticator = null;
 		if (mailAccount.isAuth()) {
 			authenticator = new UserPassAuthenticator(mailAccount.getUser(), mailAccount.getPass());
@@ -386,7 +386,7 @@ public class MailUtil {
 	 * @since 4.6.3
 	 */
 	private static String send(MailAccount mailAccount, boolean useGlobalSession, Collection<String> tos, Collection<String> ccs, Collection<String> bccs, String subject, String content,
-	                           Map<String, InputStream> imageMap, boolean isHtml, File... files) {
+							   Map<String, InputStream> imageMap, boolean isHtml, File... files) {
 		final Mail mail = Mail.create(mailAccount).setUseGlobalSession(useGlobalSession);
 
 		// 可选抄送人